Populate 'ignored_fields' with ignored fields
Currently 'ignored_fields' 'HostParams's field is a bit misleading since
it contains config's fields that weren't successfully parsed. As a
result, it's not possible to access config fields that were properly parsed
but that aren't supported yet.
In order to fix this, an additional 'unparsed_fields' field was
introduced into 'HostParams' to contain config fields that failed to
parse, while 'ignored_fields' now contains fields that aren't supported.